Fix issue with Font Awesome icons not showing correctly
The previous implementation required the Font Awesome font to be
installed locally for the icons to be displayed (empty boxes would be
shown otherwise). This is fixed by switching to the more extensive
`all.min.css` file. I also decided to use the file hosted online instead
of keeping a copy of the file in this repo.